% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numerical integration.R
\name{cdf}
\alias{cdf}
\title{Numerical integration with integrate using R}
\usage{
cdf(x, p, q)
}
\arguments{
\item{x}{samples of one distribution}

\item{y}{samples of one distribution}
}
\value{
the value of the Numerical integration with integrate \code{n}
}
\description{
Numerical integration with integrate using R
}
