% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer-von mises.R
\name{cramer}
\alias{cramer}
\title{Cramer-von Mises statistic using R}
\usage{
cramer(x, y)
}
\arguments{
\item{x}{samples of one distribution}

\item{y}{samples of one distribution}
}
\value{
the value of the Cramer-von Mises statistic \code{n}
}
\description{
Cramer-von Mises statistic using R
}
\examples{
\dontrun{
attach(chickwts)
x <- sort(as.vector(weight[feed == "soybean"]))
y <- sort(as.vector(weight[feed == "linseed"]))
detach(chickwts)
w2.0 <- cramer(x,y)
print(w2.0)
}
}
