% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytable.R
\name{mytable}
\alias{mytable}
\title{Table}
\usage{
mytable(x, y)
}
\arguments{
\item{x}{one object}

\item{y}{one object}
}
\value{
a table \code{n}
}
\description{
build a contingency table of the counts at each combination of factor levels with two objects
}
\examples{
\dontrun{
x<-c(4,5,6,7,8)
y<-c(1,2,3,3,4)
mytable(x,y)
}
}
